+++ backend/src/migrations.rs
@@ -280,6 +280,10 @@ pub enum SchemaVersion {
     CreatePltTokenAndEventTables,
     #[display("0040: Alter PLT events add event_timestamp and index")]
     AlterPltEventsAddEventTimestampAndIndex,
+    #[display("0041: Re-Add the index `baker_related_tx_idx` that was dropped in mistake due to column drop")]
+    ReAddBakerRelatedTransactionsIndex,
+    #[display("0042: Alter PLT accounts add NOT NULL constraint on amount and decimal")]
+    AlterPltAccountsAddNotNullConstraint,
 }
 impl SchemaVersion {
     /// The minimum supported database schema version for the API.
@@ -288,7 +292,7 @@ impl SchemaVersion {
     pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion =
         SchemaVersion::CreatePltTokenAndEventTables;
     /// The latest known version of the schema.
-    const LATEST: SchemaVersion = SchemaVersion::AlterPltEventsAddEventTimestampAndIndex;
+    const LATEST: SchemaVersion = SchemaVersion::AlterPltAccountsAddNotNullConstraint;
 
     /// Parse version number into a database schema version.
     /// None if the version is unknown.
@@ -350,6 +354,8 @@ impl SchemaVersion {
             SchemaVersion::BakerApyQueryUpdateProtectAgainstOverflow => false,
             SchemaVersion::CreatePltTokenAndEventTables => false,
             SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => false,
+            SchemaVersion::ReAddBakerRelatedTransactionsIndex => false,
+            SchemaVersion::AlterPltAccountsAddNotNullConstraint => false,
         }
     }
 
@@ -400,6 +406,8 @@ impl SchemaVersion {
             SchemaVersion::BakerApyQueryUpdateProtectAgainstOverflow => false,
             SchemaVersion::CreatePltTokenAndEventTables => false,
             SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => false,
+            SchemaVersion::ReAddBakerRelatedTransactionsIndex => false,
+            SchemaVersion::AlterPltAccountsAddNotNullConstraint => false,
         }
     }
 
@@ -674,8 +682,23 @@ impl SchemaVersion {
                     .await?;
                 SchemaVersion::AlterPltEventsAddEventTimestampAndIndex
             }
-
-            SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => unimplemented!(
+            SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => {
+                tx.as_mut()
+                    .execute(sqlx::raw_sql(include_str!(
+                        "./migrations/m0041_re_add_index_baker_related_transactions.sql"
+                    )))
+                    .await?;
+                SchemaVersion::ReAddBakerRelatedTransactionsIndex
+            }
+            SchemaVersion::ReAddBakerRelatedTransactionsIndex => {
+                tx.as_mut()
+                    .execute(sqlx::raw_sql(include_str!(
+                        "./migrations/m0042-plt-account-alter-add-not-null-constraint.sql"
+                    )))
+                    .await?;
+                SchemaVersion::AlterPltAccountsAddNotNullConstraint
+            }
+            SchemaVersion::AlterPltAccountsAddNotNullConstraint => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()
             ),